#430 - Generalize agent ping -> instruction mechanism.
This is an attempt to generalize the mechanism used to cancel a job. Every ping of the agent
to the server used to return an 'Instruction' to the agent. At this point, the instruction
was always to either cancel the running job or to not cancel it.

Due to this change, the instructions are routed through a router, giving a chance to other
handlers to handle an instruction.