#430 - Remove potential race condition.
There was a potential of a race between the check of cachedChecksum == null and the return.
Now, since the copyOfCachedChecksum points to the reference of cachedChecksum, it cannot change
between the two calls. Of course, there is a possibility of the value being stale now, but in
this case, that is alright. The next ping from the agent will get the right value.