#743 - List toggles UI.
The UI is really bad. This is just hooking up everything, so that it can
be improved soon.

Want to try and push more towards the Javascript, rather than writing
untestable HTML in ERB. Introduced handlebars.js for templating: Makes
it so much easier, than hand binding values to DOM.