WIP - Allow only certain ciphers.
The server already restricts the ciphers available for connecting agents. In some cases,
when Java versions are different, the SSL handshake fails to pick one of those ciphers,
and Jetty 6 fails to accept an upload from the agent, and fails quietly. This change
is about restricting the client connection to the three ciphers.

This does not reduce security. That has already been done by restricting ciphers on the
server side.