Fix bug in PipelineGroupModel#add
The add() method was creating a new PipelineModel object after
deconstructing (or getting all the values it thought it needed) from the
passed in PipelineModel object. However, canAdminister was not part of
the constructor. So, it was not copied over and always defaulted to
false. So, if you created a PipelineModel with canAdminister set to true
and then called groupModel.add(), it would add into its models a
PipelineModel with canAdminister set to false.

Now, a copy constructor is used, which copies over canAdminister and the
test uses reflectionEquals to do a deep equality check, to ensure that
any new fields added here are copied over too.